refactor(graphs): rename severity-utils to utils and update imports
Generalize utility file name from severity-utils.ts to utils.ts for better
scalability as more utility functions are added. Update import references
across BarChart, HorizontalBarChart, and ScatterPlot components.
