+++ ui/components/graphs/shared/utils.ts
